# docs/openapi.yml
openapi: 3.1.0
info:
  version: 1.0.0
  title: Screenlog
  contact:
    name: Bálint Kozák
    email: kozak.balint15@gmail.com
  description: |
    This is the api documentation for the Screenlog project.
...
servers:
  - url: "http://localhost:6969/v1"
    description: Local server
paths:
...
  "/watched":
    get:
      tags:
        - Watched
      summary: Get the watched list
      description: |
        This endpoint is used to get the watched list of the current user.
      operationId: GetWatchedHandler
      security:
        - main_auth: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchedResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  "/watched/movies/{id}":
    get:
      tags:
        - Watched
      summary: Check if a movie is in the watched list
      description: |
        This endpoint is used to check if a movie is in the watched list of the current user.
      operationId: CheckMovieInWatchedHandler
      security:
        - main_auth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The movie id
          schema:
            type: integer
            format: int64
            examples:
              - 278
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InWatched"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    post:
      tags:
        - Watched
      summary: Add a movie to the watched list
      description: |
        This endpoint is used to add a movie to the watched list of the current user.
      operationId: AddMovieToWatchedHandler
      security:
        - main_auth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The movie id
          schema:
            type: integer
            format: int64
            examples:
              - 278
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      tags:
        - Watched
      summary: Remove a movie from the watched list
      description: |
        This endpoint is used to remove a movie from the watched list of the current user.
      operationId: RemoveMovieFromWatchedHandler
      security:
        - main_auth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The movie id
          schema:
            type: integer
            format: int64
            examples:
              - 278
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  "/watched/tv/{id}":
    get:
      tags:
        - Watched
      summary: Check if a TV show is in the watched list
      description: |
        This endpoint is used to check if a TV show is in the watched list of the current user.
      operationId: CheckTvInWatchedHandler
      security:
        - main_auth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The TV show id
          schema:
            type: integer
            format: int64
            examples:
              - 1399
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchedTvResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    post:
      tags:
        - Watched
      summary: Add a TV show to the watched list
      description: |
        This endpoint is used to add a TV show to the watched list of the current user.
      operationId: AddTvToWatchedHandler
      security:
        - main_auth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The TV show id
          schema:
            type: integer
            format: int64
            examples:
              - 1399
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      tags:
        - Watched
      summary: Remove a TV show from the watched list
      description: |
        This endpoint is used to remove a TV show from the watched list of the current user.
      operationId: RemoveTvFromWatchedHandler
      security:
        - main_auth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The TV show id
          schema:
            type: integer
            format: int64
            examples:
              - 1399
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Internal server error
  "/watched/tv/{id}/season/{season_number}":
    get:
      tags:
        - Watched
      summary: Check if a TV show season episodes are in the watched list
      description: |
        This endpoint is used to check if a TV show season episodes are in the watched list of the current user.
      operationId: CheckTvSeasonInWatchedHandler
      security:
        - main_auth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The TV show id
          schema:
            type: integer
            format: int64
            examples:
              - 1399
        - name: season_number
          in: path
          required: true
          description: The season number
          schema:
            type: integer
            format: int64
            examples:
              - 1
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchedTvSeasonResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    post:
      tags:
        - Watched
      summary: Add a TV show season to the watched list
      description: |
        This endpoint is used to add a TV show season to the watched list of the current user.
      operationId: AddTvSeasonToWatchedHandler
      security:
        - main_auth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The TV show id
          schema:
            type: integer
            format: int64
            examples:
              - 1399
        - name: season_number
          in: path
          required: true
          description: The season number
          schema:
            type: integer
            format: int64
            examples:
              - 1
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      tags:
        - Watched
      summary: Remove a TV show season from the watched list
      description: |
        This endpoint is used to remove a TV show season from the watched list of the current user.
      operationId: RemoveTvSeasonFromWatchedHandler
      security:
        - main_auth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The TV show id
          schema:
            type: integer
            format: int64
            examples:
              - 1399
        - name: season_number
          in: path
          required: true
          description: The season number
          schema:
            type: integer
            format: int64
            examples:
              - 1
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  "/watched/tv/{id}/season/{season_number}/episode/{episode_number}":
    get:
      tags:
        - Watched
      summary: Check if a TV show episode is in the watched list
      description: |
        This endpoint is used to check if a TV show episode is in the watched list of the current user.
      operationId: CheckTvEpisodeInWatchedHandler
      security:
        - main_auth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The TV show id
          schema:
            type: integer
            format: int64
            examples:
              - 1399
        - name: season_number
          in: path
          required: true
          description: The season number
          schema:
            type: integer
            format: int64
            examples:
              - 1
        - name: episode_number
          in: path
          required: true
          description: The episode number
          schema:
            type: integer
            format: int64
            examples:
              - 1
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InWatched"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    post:
      tags:
        - Watched
      summary: Add a TV show episode to the watched list
      description: |
        This endpoint is used to add a TV show episode to the watched list of the current user.
      operationId: AddTvEpisodeToWatchedHandler
      security:
        - main_auth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The TV show id
          schema:
            type: integer
            format: int64
            examples:
              - 1399
        - name: season_number
          in: path
          required: true
          description: The season number
          schema:
            type: integer
            format: int64
            examples:
              - 1
        - name: episode_number
          in: path
          required: true
          description: The episode number
          schema:
            type: integer
            format: int64
            examples:
              - 1
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    delete:
      tags:
        - Watched
      summary: Remove a TV show episode from the watched list
      description: |
        This endpoint is used to remove a TV show episode from the watched list of the current user.
      operationId: RemoveTvEpisodeFromWatchedHandler
      security:
        - main_auth: []
      parameters:
        - name: id
          in: path
          required: true
          description: The TV show id
          schema:
            type: integer
            format: int64
            examples:
              - 1399
        - name: season_number
          in: path
          required: true
          description: The season number
          schema:
            type: integer
            format: int64
            examples:
              - 1
        - name: episode_number
          in: path
          required: true
          description: The episode number
          schema:
            type: integer
            format: int64
            examples:
              - 1
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
components:
  securitySchemes:
    main_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SearchMovie:
      type: object
      required:
        - id
        - title
        - overview
        - poster_url
        - release_date
        - vote_average
        - popularity
      properties:
        id:
          type: integer
          format: int64
          examples:
            - 278
        title:
          type: string
          examples:
            - "The Shawshank Redemption"
        overview:
          type: string
          examples:
            - "Framed in the 1940s for the double murder of his wife and her lover, upstanding banker Andy Dufresne begins a new life at the Shawshank prison, where he puts his accounting skills to work for an amoral warden. During his long stretch in prison, Dufresne comes to be admired by the other inmates -- including an older prisoner named Red -- for his integrity and unquenchable sense of hope."
        poster_url:
          type: string
          examples:
            - "https://image.tmdb.org/t/p/w92/9cqNxx0GxF0bflZmeSMuL5tnGzr.jpg"
        release_date:
          type: string
          examples:
            - "1994-09-23"
        vote_average:
          type: number
          format: float
          examples:
            - 8.3
        popularity:
          type: number
          format: float
          examples:
            - 186.654
...
    SearchTv:
      type: object
      required:
        - id
        - title
        - overview
        - poster_url
        - release_date
        - vote_average
        - popularity
      properties:
        id:
          type: integer
          format: int64
          examples:
            - 1399
        title:
          type: string
          examples:
            - "Game of Thrones"
        overview:
          type: string
          examples:
            - "Seven noble families fight for control of the mythical land of Westeros. Friction between the houses leads to full-scale war. All while a very ancient evil awakens in the farthest north. Amidst the war, a neglected military order of misfits, the Night's Watch, is all that stands between the realms of men and icy horrors beyond."
        poster_url:
          type: string
          examples:
            - "https://image.tmdb.org/t/p/w92/1XS1oqL89opfnbLl8WnZY1O1uJx.jpg"
        release_date:
          type: string
          examples:
            - "2011-04-17"
        vote_average:
          type: number
          format: float
          examples:
            - 8.4
        popularity:
          type: number
          format: float
          examples:
            - 749.726
...
    WatchedResponse:
      type: object
      required:
        - watched
      properties:
        watched:
          type: object
          $ref: "#/components/schemas/Watched"
    Watched:
      type: object
      required:
        - movies
        - tv
      properties:
        movies:
          type: array
          items:
            $ref: "#/components/schemas/SearchMovie"
        tv:
          type: array
          items:
            $ref: "#/components/schemas/SearchTv"
    InWatched:
      type: object
      required:
        - in_watched
      properties:
        in_watched:
          type: boolean
          examples:
            - true
    WatchedTvResponse:
      type: object
      required:
        - watched_tv
      properties:
        watched_tv:
          type: object
          $ref: "#/components/schemas/WatchedTv"
    WatchedTv:
      type: object
      required:
        - id
        - status
        - seasons
        - progress
      properties:
        id:
          type: integer
          format: int64
          examples:
            - 1399
        status:
          type: string
          enum: ["watched", "in_progress", "not_watched"]
        progress:
          type: number
          format: double
          examples:
            - 100
            - 33.33
        seasons:
          type: array
          items:
            $ref: "#/components/schemas/WatchedTvSeason"
    WatchedTvSeasonResponse:
      type: object
      required:
        - watched_season
      properties:
        watched_season:
          type: object
          $ref: "#/components/schemas/WatchedTvSeason"
    WatchedTvSeason:
      type: object
      required:
        - season_number
        - status
        - episodes
      properties:
        season_number:
          type: integer
          examples:
            - 1
        status:
          type: string
          enum: ["watched", "in_progress", "not_watched"]
        episodes:
          type: array
          items:
            type: integer
            examples:
              - 1
              - 2
              - 3
    Message:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          examples:
            - "Internal server error"
            - "Bad request"
            - "Unauthorized"
            - "Forbidden"
            - "Not found"
            - "Movie added to watchlist"
            - "Movie removed from watchlist"
            - "TV show added to watchlist"
            - "TV show removed from watchlist"
            - "TV show season added to watchlist"
            - "TV show season removed from watchlist"
            - "TV show episode added to watchlist"
            - "TV show episode removed from watchlist"